se mueve verifyToken a auth
